Create cpp and python algorithms for cutting image
Created implementation of cutting shapes in images for python and cpp.
Cpp implementation uses class type shapes and python hard coded ones.

Both algorithms take original photo, position and size as a starting parameters and as a result give the same photo with the shape cut. The position and size
differs from the starting parameters.